backend/requirements.txt
Flask
Flask-CORS
python-jose
